requirements.txt
pyrogram==2.0.80
TgCrypto
motor
dnspython
hachoir
Pillow
aiohttp
pytz
humanize
pyromod
python-dotenv
ffmpeg-python
